+++ resources/js/app.js
@@ -1,6 +1,7 @@
 import { createApp, h } from 'vue'
 import { InertiaProgress } from '@inertiajs/progress'
 import { createInertiaApp } from '@inertiajs/inertia-vue3'
+import PrimeVue from 'primevue/config';
 
 InertiaProgress.init()
 
@@ -10,6 +11,7 @@ createInertiaApp({
     setup({ el, App, props, plugin }) {
         createApp({ render: () => h(App, props) })
             .use(plugin)
+            .use(PrimeVue, {inputStyle: 'filled', ripple: true})
             .mount(el)
     },
 });