+++ resources/js/components/auth/Panel.vue
@@ -1,0 +1,13 @@
+<template>
+    <div class="bg-white">
+        <i class="pi pi-sign-out"></i>
+    </div>
+</template>
+
+<script>
+export default {
+    props: {
+        authOpen: Boolean,
+    },
+};
+</script>